Fixed overlapping colliders drag bug
When the portal was overlapping with another collider, the drag was not working. it was because of the onmouse methods that are not able to separate the multiple colliders. so, i changed the input to update method and used an overlappoint to detect clicking on the object with a layermask.